+++ lib/shoulda/matchers/active_record/association_matcher.rb
@@ -162,6 +162,24 @@ module Shoulda
       #       should belong_to(:world).dependent(:destroy)
       #     end
       #
+      # To assert that *any* `:dependent` option was specified, use `true`:
+      #
+      #     # RSpec
+      #     describe Person do
+      #       it { should belong_to(:world).dependent(true) }
+      #     end
+      #
+      # To assert that *no* `:dependent` option was specified, use `false`:
+      #
+      #     class Person < ActiveRecord::Base
+      #       belongs_to :company
+      #     end
+      #
+      #     # RSpec
+      #     describe Person do
+      #       it { should belong_to(:company).dependent(false) }
+      #     end
+      #
       # ##### counter_cache
       #
       # Use `counter_cache` to assert that the `:counter_cache` option was
@@ -993,7 +1011,7 @@ module Shoulda
         end
 
         def macro_supports_primary_key?
-          macro == :belongs_to || 
+          macro == :belongs_to ||
             ([:has_many, :has_one].include?(macro) && !through?)
         end